Add support for uploading HAR files to `upload` (#13)
This adds support for uploading HAR files to the `upload` command. The old `--service` and `--name` options are now deprecated in favour of `--dest`, which takes in Akita URIs.

When uploading traces, the user may provide multiple files to upload to a single trace. Plugins and removal of third-party trackers is supported. If the specified trace already exists, the user must specify `--append` to modify the trace. If `--append` is used when the trace doesn't yet exist, the trace is created and a warning is printed to stderr.