Revamped daemon APIs
When polling services for new traces, middleware and the daemon now send
the set of active traces that they are aware of, and are notified of any
new traces that become active.

Each trace now has a new endpoint for polling for the end of that trace.

The heartbeat now includes the daemon's name. In the future, we can
expand this to identify the middleware logging for each active trace.