Revamped daemon APIs again
Polling for the activation and deactivation of traces have been merged
into a single endpoint, both in the daemon's and the cloud's APIs.
Clients of the daemon and cloud now provide a set of traces they think
are active for each service. The server responds when the set becomes
different from what is in the request.