Add --cluster as an alias for --service
We're moving away from a per-service model, where every trace is associated
with a single service, to a mixed model, where traces may contain events from
different services.

As an intermediate step, this PR introduces --cluster as an alias for
--service.  Under the hood, both services and clusters are handled identically.
However, the --cluster flag will help avoid confusion in our documentation when
deploying Akita for monitoring in staging/production environments.