Only sample TCP connections when rate limiter samples witnesses
Not perfect, since the number of TCP connections we see could very well
be larger than the number of HTTP requests. Doing it right appears to
require witnesses and connections to have independent sampling
intervals, but the rate limiter is not set up very well to support this.

We can revisit this if needed.