Detect when interface failures are due to arch
When running a dockerized version of the Akita agent on a host with a different
architecture than the container, the agent fails to read any network interfaces
with a "Function not implemented" error.

This PR detects these errors and prints tailored warning/error messages
instructing users to try using an agent binary/pulling an image that matches
their host architecture.