Upload metadata with initial client telemetry (#181)
Call a new backend API to report initial client telemetry, and include the following additional metadata:
* CLI version
* CLI target architecture
* Whether the CLI was running in a Docker container distributed by Akita
* Whether the CLI was running in a Docker container running on Docker Desktop

The main goal is to detect when customers are running the Akita CLI in Docker Desktop with their service running on the host.  Because Docker Desktop (e.g. on macOS) doesn't support `--network host`, users in this case won't capture any service traffic.  Collecting this metadata lets us display a nicer error in the Akita app.