Add `ecs cf-fragment` and `ecs task-def` subcommands
Added `ecs cf-fragment` subcommand for printing an AWS CloudFormation
fragment for adding the agent to ECS.

Added `ecs task-def` subcommand for printing a task definition that can
be added to an ECS cluster to run the agent as a daemon in
host-networking mode on every EC2 instance in the cluster.

Unfortunately, the types defined by AWS do not have `json:",omitempty"`
on their fields, so the output is not as compact as it could be.