[POA-168] Clean up authentication messages (#267)
In this PR, we have updated some of the error messages shown to users in
case of errors.
* Remove mention of login command with postman-insights-agent
* Remove mention of contacting the postman support team if a user is
using --collection flag
* The 3rd message mentioned in the JIRA ticket has already been fixed in
earlier releases.
* Don't show the raw response body to the user for external API call
errors; instead, show a generic message for unhandled cases.


**PS**: To show the user the proper message, we will need to follow a
standard error response model, which will involve changes in both the
Agent and Superstar services. Otherwise, we will face issues in mapping
the response to the proper Golang object.